Move auth provides to their own files
Prevent cyclical dependencies
